HMS/requirements.txt
django>=5.0
psycopg2-binary
python-dotenv
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
requests
Pillow
